fix: simulator root directory should be absolute
When set to a relative path, parsing of gamelist.xml would fail because
we try to canonicalize a path, resulting in an absolute path. Then, when
we look at the filesystem, we get relative paths. This causes the
deduplication logic to keep both versions of the path, so the changes in
gamelist.xml is not saved to the database. Because the gamelist.xml is
only parsed when changed, we don't parse it the next time and lose all
changes.